refactor: file namings, getter for DatesItem
