+++ apps/anime-service/src/modules/dates/items/future-dates.item.ts
@@ -1,4 +1,4 @@
-import { DatesItem } from './dates-item';
+import { DatesItem } from './dates-item.abstract';
 
 export class FutureDates extends DatesItem {
   getDates(startDate: Date, endDate: Date) {
