+++ apps/anime-service/src/modules/dates/items/present-dates.item.ts
@@ -1,0 +1,9 @@
+import { DatesItem } from './dates-item.abstract';
+
+export class PresentDates extends DatesItem {
+  getDates(startDate: Date, endDate: Date) {
+    const currentDate = this.getCurrentDate();
+
+    return { startDate: startDate ?? currentDate, endDate: null };
+  }
+}
